> Completed Thread Safe implementation of the export process
